make sure menu values are not limited to 8bits

+++ app/inc/system.h
@@ -61,7 +61,7 @@
 
 // system menu callbacks
 void system_lock_comm_serial(uint8_t bussy);
-void system_update_menu_value(uint8_t item_ID, uint8_t value);
+void system_update_menu_value(uint8_t item_ID, uint16_t value);
 uint8_t system_get_current_profile(void);
 uint8_t master_volume(int event, int step);
 void system_pedalboard_cb(void *arg, int event);
